backend/requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
boto3==1.21.32
botocore==1.24.32
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.2
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
django-environ==0.8.1
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
filelock==3.6.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
jmespath==1.0.0
outcome==1.1.0
pipenv==2022.1.8
platformdirs==2.5.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
virtualenv==20.13.4
virtualenv-clone==0.5.7
whitenoise==6.0.0
wsproto==1.1.0
